Move manifest-updater storage code to key-rotator manifest package.
No code changes yet, this commit just moves the existing code & updates
references to it.
