+++ deploy-tool/main.go
@@ -25,7 +25,7 @@ import (
 
 	"github.com/abetterinternet/prio-server/deploy-tool/key"
 	"github.com/abetterinternet/prio-server/key-rotator/manifest"
-	manifest_storage "github.com/abetterinternet/prio-server/manifest-updater/manifest"
+	"github.com/abetterinternet/prio-server/key-rotator/storage"
 )
 
 // This tool consumes the output of `terraform apply`, generating keys and then
@@ -51,7 +51,7 @@ type SpecificManifestWrapper struct {
 // variables this program is interested in.
 type TerraformOutput struct {
 	ManifestBucket struct {
-		Value manifest_storage.Bucket
+		Value storage.Bucket
 	} `json:"manifest_bucket"`
 	SpecificManifests struct {
 		Value map[string]SpecificManifestWrapper
@@ -130,7 +130,7 @@ func generateAndDeployKeyPair(
 func setupTestEnvironment(
 	k8sSecretsClientGetter k8scorev1.SecretsGetter,
 	ingestor *SingletonIngestor,
-	manifestWriter manifest_storage.Writer,
+	manifestWriter storage.Writer,
 ) error {
 	batchSigningPublicKey, err := createBatchSigningPublicKey(
 		k8sSecretsClientGetter.Secrets(ingestor.TesterKubernetesNamespace),
@@ -191,7 +191,7 @@ func createManifest(
 	k8sSecretsClientGetter k8scorev1.SecretsGetter,
 	dataShareProcessorName string,
 	manifestWrapper *SpecificManifestWrapper,
-	manifestWriter manifest_storage.Writer,
+	manifestWriter storage.Writer,
 	packetEncryptionKeyCSRs manifest.PacketEncryptionKeyCSRs,
 ) error {
 	k8sSecretsClient := k8sSecretsClientGetter.Secrets(manifestWrapper.KubernetesNamespace)
@@ -311,7 +311,7 @@ func backupKeys(
 func createManifests(
 	k8sSecretsClientGetter k8scorev1.SecretsGetter,
 	specificManifests map[string]SpecificManifestWrapper,
-	manifestStorage manifest_storage.Storage,
+	manifestStorage storage.Storage,
 ) error {
 	// Iterate over all specific manifests described by TF output so we can
 	// record any packet encryption key CSRs that have already been created. We
@@ -409,13 +409,13 @@ func main() {
 		log.Fatalf("%s", err)
 	}
 
-	manifestStorage, err := manifest_storage.NewStorage(&terraformOutput.ManifestBucket.Value)
+	manifestStorage, err := storage.NewStorage(&terraformOutput.ManifestBucket.Value)
 	if err != nil {
 		log.Fatalf("%s", err)
 	}
 
 	if terraformOutput.HasTestEnvironment.Value && terraformOutput.SingletonIngestor.Value != nil {
-		globalManifestStorage, err := manifest_storage.NewStorage(&manifest_storage.Bucket{
+		globalManifestStorage, err := storage.NewStorage(&storage.Bucket{
 			URL:        terraformOutput.ManifestBucket.Value.URL,
 			KeyPrefix:  "singleton-ingestor",
 			AWSRegion:  terraformOutput.ManifestBucket.Value.AWSRegion,
