+++ key-rotator/go.mod
@@ -2,4 +2,28 @@ module github.com/abetterinternet/prio-server/key-rotator
 
 go 1.17
 
-require github.com/google/go-cmp v0.5.6
+require (
+	cloud.google.com/go/storage v1.18.2
+	github.com/aws/aws-sdk-go v1.41.16
+	github.com/google/go-cmp v0.5.6
+	github.com/rs/zerolog v1.26.0
+)
+
+require (
+	cloud.google.com/go v0.97.0 // indirect
+	github.com/golang/groupcache v0.0.0-20200121045136-8c9f03a8e57e // indirect
+	github.com/golang/protobuf v1.5.2 // indirect
+	github.com/googleapis/gax-go/v2 v2.1.1 // indirect
+	github.com/jmespath/go-jmespath v0.4.0 // indirect
+	go.opencensus.io v0.23.0 // indirect
+	golang.org/x/net v0.0.0-20210805182204-aaa1db679c0d // indirect
+	golang.org/x/oauth2 v0.0.0-20211005180243-6b3c2da341f1 // indirect
+	golang.org/x/sys v0.0.0-20210917161153-d61c044b1678 // indirect
+	golang.org/x/text v0.3.6 // indirect
+	golang.org/x/xerrors v0.0.0-20200804184101-5ec99f83aff1 // indirect
+	google.golang.org/api v0.58.0 // indirect
+	google.golang.org/appengine v1.6.7 // indirect
+	google.golang.org/genproto v0.0.0-20211016002631-37fc39342514 // indirect
+	google.golang.org/grpc v1.40.0 // indirect
+	google.golang.org/protobuf v1.27.1 // indirect
+)
