+++ key-rotator/storage/manifest.go
@@ -1,4 +1,4 @@
-package manifest
+package storage
 
 import (
 	"bytes"
